requirements.txt
fastapi
uvicorn
scikit-learn
numpy
google-genai
python-multipart
pydantic
